Copy aseg.presurf.mgz
Due to updates in the recon-all processing stream, the volume called aseg.presurf.mgz should be edited when using FreeSurfer version 6.0 or greater. For versions 5.3 and less, directly edit the aseg.mgz volume. In version 6.0, the aseg.mgz is cleaned up using the surfaces, so that the cortical gray matter in the aseg volume follows the surface boundaries. As such, any edits on the aseg.mgz volume aren't taken into account in the earlier stages of recon-all. The aseg.presurf.mgz volume is the "unclean" version of the aseg, and is equivalent to the aseg.mgz volume in previous FS versions (5.3 or less). This cleanup step is only to improve the visual appearance of the aseg volume; in all versions of FreeSurfer, statistics for cortical gray matter and white matter are calculated using the surfaces.